pyproject.toml
[project]
name = "market-analyzer"
version = "0.1.0"
description = "Data-driven market analyzer using a hierarchical supervisor pattern with researcher and coder agents"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "python-dotenv>=1.0.0",
    "langchain>=0.3.0",
    "langchain-openai>=0.3.0",
    "langchain-community>=0.3.0",
    "langchain-core>=0.3.0",
    "langgraph>=0.2.0",
    "langgraph-supervisor>=0.0.1",
    "duckduckgo-search>=6.0.0",
    "streamlit>=1.40.0",
    "ddgs>=9.12.1",
]
